Shift floor up
Now that that issue has been fixed (41e51f5) there is no need for a
1-pixel buffer between the player (even that's a stretch) and floor.
Obviously this is a git repository. I didn't make small, incremental
commits for the code I already had like I wanted to, to record all I
learned so far. Still there's enough from the beginning to here.
Even using a velocity of 1 can be too fast, so keep track of these
variables with more detail. Either change type to some floating point
or use integers but divide by a constant (i.e. y is 0-960 and divide by
4 to get a position on the grid and screen).
Furthermore speed is attached to the delay after each iteration of the
game loop. Make character speed and game loop rate not contingent on
one another. Look in to any form of timer.