Sprite with volume
Add a drawRect method back from 75fb6f1. I could have made it a helper
method in main.c but it seems like something that will be useful
elsewhere, so I added it to gfx.
Use drawRect to display a square character instead of a point. This
displays the problem with graphics mode 3 (160px by 240px, 32768
colors). Some flickering is visible when frame is drawn and the game
loop has not finished erasing or drawing the sprite.
https://cloud.githubusercontent.com/assets/5026621/21627981/d5d2c7aa-d1e
9-11e6-805d-25b1144d4717.png
Also now drawing the floor where it actually is and the sprite sits on
top of it.