Fix horizontal screen wrap
Before everything appeared to work perfectly because of VRAM's
sequential nature in bitmap graphics mode 3. However, this does cause
the floor to be overwritten. Use a simple modulo operator to wrap
around pixels past the left edge of the screen.
Note the floor could be redrawn each game loop iteration, but this is
still a fix.