Page flipping with graphics mode 4
I wanted to try out graphics mode 4 but it turns out d83f98d is just so
much more simple. The main advantage to using mode 4 is that you can
take as long as you like to draw the next frame before flipping it to
be visible. The palette may also be useful for quickly changing the
color of large areas. In this use case the drawing can easily be done
in within vblank. For more complicated (or even 3D) graphics this could
be useful.
Note: this is not double-buffering! There is no "blit" step where the frame
is copied from the back buffer to the primary surface. Instead,
the video pointer is flipped between two buffers.
Nice graphical explanation: https://docs.oracle.com/javase/tutorial/extra/fullscreen/doublebuf.html