Platform code improvements
Before the character was overwriting the platform because you could
jump upwards through it and it wasn't ever re-drawn.
The character was flickering a bit and felt like it was moving too
fast, increased the delay at the end of the game loop to fix both.
Cleared player's y-velocity when in contact with a platform so it
wouldn't build up and make them immediately fall to the ground when
leaving the platform horizontally.